GCC 12, Ice Lake processor.

Main branch (we repeat three times)

```
BasicBench_AdaURL_aggregator_href   23115321 ns     23079817 ns           30 GHz=3.18789 cycle/byte=8.45902 cycles/url=734.744 instructions/byte=25.1299 instructions/cycle=2.97078 instructions/ns=9.47052 instructions/url=2.18276k ns/url=230.48 speed=376.437M/s time/byte=2.65649ns time/url=230.74ns url/s=4.33387M/s
BasicBench_AdaURL_aggregator_href   23186567 ns     23152848 ns           30 GHz=3.18677 cycle/byte=8.49247 cycles/url=737.649 instructions/byte=25.1299 instructions/cycle=2.95908 instructions/ns=9.42992 instructions/url=2.18276k ns/url=231.472 speed=375.249M/s time/byte=2.66489ns time/url=231.471ns url/s=4.3202M/s
BasicBench_AdaURL_aggregator_href   23083173 ns     23045943 ns           30 GHz=3.18829 cycle/byte=8.4514 cycles/url=734.082 instructions/byte=25.1299 instructions/cycle=2.97346 instructions/ns=9.48026 instructions/url=2.18276k ns/url=230.243 speed=376.99M/s time/byte=2.65259ns time/url=230.402ns url/s=4.34024M/s
```

This PR:

```
BasicBench_AdaURL_aggregator_href   22530622 ns     22497930 ns           31 GHz=3.18865 cycle/byte=8.24613 cycles/url=716.252 instructions/byte=24.7497 instructions/cycle=3.00137 instructions/ns=9.57031 instructions/url=2.14974k ns/url=224.626 speed=386.173M/s time/byte=2.58951ns time/url=224.923ns url/s=4.44596M/s
BasicBench_AdaURL_aggregator_href   22497005 ns     22462658 ns           31 GHz=3.18802 cycle/byte=8.2335 cycles/url=715.155 instructions/byte=24.7497 instructions/cycle=3.00598 instructions/ns=9.58313 instructions/url=2.14974k ns/url=224.326 speed=386.779M/s time/byte=2.58545ns time/url=224.57ns url/s=4.45295M/s
BasicBench_AdaURL_aggregator_href   22489804 ns     22456437 ns           30 GHz=3.18553 cycle/byte=8.23304 cycles/url=715.115 instructions/byte=24.7497 instructions/cycle=3.00614 instructions/ns=9.57615 instructions/url=2.14974k ns/url=224.489 speed=386.886M/s time/byte=2.58474ns time/url=224.508ns url/s=4.45418M/s
```

PR https://github.com/ada-url/ada/pull/548

```
BasicBench_AdaURL_aggregator_href   22163649 ns     22133069 ns           32 GHz=3.18781 cycle/byte=8.12625 cycles/url=705.84 instructions/byte=24.0062 instructions/cycle=2.95415 instructions/ns=9.41728 instructions/url=2.08516k ns/url=221.418 speed=392.539M/s time/byte=2.54752ns time/url=221.275ns url/s=4.51926M/s
BasicBench_AdaURL_aggregator_href   22083116 ns     22051597 ns           32 GHz=3.1877 cycle/byte=8.10477 cycles/url=703.974 instructions/byte=24.0062 instructions/cycle=2.96198 instructions/ns=9.4419 instructions/url=2.08516k ns/url=220.841 speed=393.989M/s time/byte=2.53814ns time/url=220.461ns url/s=4.53595M/s
BasicBench_AdaURL_aggregator_href   22197996 ns     22163061 ns           32 GHz=3.1873 cycle/byte=8.148 cycles/url=707.729 instructions/byte=24.0062 instructions/cycle=2.94627 instructions/ns=9.39064 instructions/url=2.08516k ns/url=222.047 speed=392.008M/s time/byte=2.55097ns time/url=221.575ns url/s=4.51314M/s
```

What is apparent is that there is very little variation from run to run. I use systems configured for benchmarking.



PR https://github.com/ada-url/ada/pull/548 appears to be superior to this table-based PR, at least on this system and in this benchmark.