@the-moisrex 2023-11-01
Results are consistent with mine as well.