@lemire 2023-11-01
@the-moisrex  I have integrated your approach as a fallback case (when SIMD is not available). See other PR. I added credit to you.